Adding commandline interface for migration
