+++ .gitignore
@@ -7,3 +7,4 @@ images/
 *.so
 *.db
 *.zip
+module-migration
