+++ README.md
@@ -1,0 +1,13 @@
+# README
+
+### How to build
+
+```bash
+go build -o module-migration
+```
+
+### How to run
+
+```bash
+./module-migration migrate
+```
