+++ go.mod
@@ -2,4 +2,12 @@ module github.com/NubeIO/module-migration
 
 go 1.18
 
-require github.com/mattn/go-sqlite3 v1.14.22
+require (
+	github.com/mattn/go-sqlite3 v1.14.22
+	github.com/spf13/cobra v1.8.0
+)
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
